+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Core Public (`core-public`) - Changelog
 
+## Unreleased
+
+* Added - testing support for automated testing using Semaphore
+* Added - testing support for manual testing, using local and remote environments
 ## 1.0.0 - May 2015
 
 * Removes support for modifying the sudoers file, ensure you are using a compatible bootstrap role and base system